refactor: clean up and finalize component name
