+++ src/components/centerBlock/OverlayPlayer.vue
